Add test suite for date formatting
